[ALSA] Minor clean up and fixes for CS5535 audio driver
Modules: Documentation,CS5535 driver

Minor clean up and fixes for CS5535 audio driver.
Added an entry in ALSA-Configuration.txt, too.

+++ sound/pci/cs5535audio/cs5535audio.h
@@ -2,11 +2,11 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #ifndef __SOUND_CS5535AUDIO_H
 #define __SOUND_CS5535AUDIO_H
 
-#define cs_writel(cs5535au, reg, val) outl(val, (int) cs5535au->port + reg)
-#define cs_writeb(cs5535au, reg, val) outb(val, (int) cs5535au->port + reg)
-#define cs_readl(cs5535au, reg)	inl((unsigned short) (cs5535au->port + reg))
-#define cs_readw(cs5535au, reg)	inw((unsigned short) (cs5535au->port + reg))
-#define cs_readb(cs5535au, reg)	inb((unsigned short) (cs5535au->port + reg))
+#define cs_writel(cs5535au, reg, val)	outl(val, (cs5535au)->port + reg)
+#define cs_writeb(cs5535au, reg, val)	outb(val, (cs5535au)->port + reg)
+#define cs_readl(cs5535au, reg)		inl((cs5535au)->port + reg)
+#define cs_readw(cs5535au, reg)		inw((cs5535au)->port + reg)
+#define cs_readb(cs5535au, reg)		inb((cs5535au)->port + reg)
 
 #define CS5535AUDIO_MAX_DESCRIPTORS	128
